refactor: move sort logic and types to sort.service
